+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## FUTURE
+
+* [ADDED] - support `baseUrl` App config option value that contains path string eg.: '127.0.0.1/root/app/path'
+
 ## v0.10.3
 
 * [FIXED] an `Error` is throwed when a `Router` is trying to register a `Route` with non-unique route `uid`. `uid` should be unique in context of an `App`